Provider/requirements_djongo.txt
Django==4.2.16
sqlparse==0.2.4
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
dnspython==2.6.1
djangorestframework==3.14.0
django-cors-headers==4.3.1